[refactor #4] Entity 정의 위치 변경
