+++ internal/bootstrap/repository/root.go
@@ -30,12 +30,28 @@ type Repository struct {
 func SetRepository(cfg *config.Config) (*Repository, error) {
 	r := &Repository{}
 	r.DB = r.ConnectToDB(cfg)
+
+	if r.DB == nil {
+		return nil, fmt.Errorf("failed to connect to the database")
+	}
+
 	r.UsersRepository = users.SetUsersRepository(r.DB)
+	r.AIRepository = ai.SetAIRepository(r.DB)
+	r.AuthRepository = auth.SetAuthRepository(r.DB)
+	r.RegionsRepository = regions.SetRegionsRepository(r.DB)
+	r.CarpoolsRepository = carpools.SetCarpoolsRepository(r.DB)
+	r.CoursesRepository = courses.SetCoursesRepository(r.DB)
+	r.MatchingRepository = matching.SetMatchingRepository(r.DB)
 
 	return r, nil
 }
 
 func (r *Repository) ConnectToDB(cfg *config.Config) *gorm.DB {
+	if cfg.DBUser == "" || cfg.DBPassword == "" || cfg.DBHost == "" || cfg.DBPort == 0 || cfg.DBName == "" {
+		log.Println("invalid database configuration")
+		return nil
+	}
+
 	dsn := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?charset=utf8mb4&parseTime=True&loc=Local",
 		cfg.DBUser, cfg.DBPassword, cfg.DBHost, cfg.DBPort, cfg.DBName)
 
@@ -58,5 +74,6 @@ func (r *Repository) ConnectToDB(cfg *config.Config) *gorm.DB {
 		log.Fatalf("failed to ping database: %v", err)
 	}
 
+	log.Println("데이터베이스 연결 성공!!!!!!!1")
 	return db
 }
