+++ internal/types/matching.go
@@ -22,3 +22,17 @@ type Matching struct {
 	Categories     []Category     `gorm:"foreignKey:MatchingID"`
 	Likes          []MatchingLike `gorm:"foreignKey:MatchingID"`
 }
+
+type Category struct {
+	CategoryID int      `gorm:"primaryKey;autoIncrement"`
+	Name       string   `gorm:"size:50;not null"`
+	MatchingID int      `gorm:"not null"`
+	Matching   Matching `gorm:"foreignKey:MatchingID;references:MatchingID"`
+}
+
+type MatchingLike struct {
+	LikeID     int       `gorm:"primaryKey;autoIncrement"`
+	UserID     int       `gorm:"not null"`
+	MatchingID int       `gorm:"not null"`
+	LikedAt    time.Time `gorm:"autoCreateTime"`
+}
